feat: fix authentication with keycloak intergration

+++ src/main/java/com/hotelmanagement/quanlikhachsan/dto/response/RoomTypeResponse.java
@@ -1,9 +1,10 @@
 package com.hotelmanagement.quanlikhachsan.dto.response;
 
+import java.math.BigDecimal;
+
 public record RoomTypeResponse(
-        String id,
-        String name,
-        String description,
-        Double pricePerNight
-) {
+                String id,
+                String name,
+                String description,
+                BigDecimal pricePerNight) {
 }
